Setting Tables

+++ database/migrations/2018_08_01_001539_create_products_table.php
@@ -14,7 +14,7 @@ class CreateProductsTable extends Migration
     public function up()
     {
         Schema::create('products', function (Blueprint $table) {
-            $table->bigInteger('codBar');
+            $table->increments('codBar');
             $table->string('name');
             $table->double('price', 8,2);
             $table->integer('qtd');
